+++ src/ui.rs
@@ -1,0 +1,23 @@
+use skim::prelude::*;
+use std::io::Cursor;
+
+pub fn skim_pick(prompt: &str, options: Vec<String>) -> Option<String> {
+    let input = options.join("\n");
+    let prompt_str = format!("{}> ", prompt);
+    let options = SkimOptionsBuilder::default()
+        .prompt(prompt_str)
+        .multi(false)
+        .bind(vec!["esc:abort".to_string()])
+        .build()
+        .unwrap();
+
+    let item_reader = SkimItemReader::default();
+    let items = item_reader.of_bufread(Cursor::new(input));
+    let output = Skim::run_with(&options, Some(items))?;
+
+    if output.is_abort || output.selected_items.is_empty() {
+        None
+    } else {
+        Some(output.selected_items[0].output().to_string())
+    }
+}
